pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pms"
version = "0.1.0"
description = "Process Modeling System (PMS) orchestrator and ML pipelines"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.0.0",
  "numpy>=1.24.0",
  "scikit-learn>=1.3.0",
  "xgboost>=1.7.0",
  "optuna>=3.6.0",
  "keras-tuner>=1.4.7",
  "tensorflow>=2.15.0; platform_machine != 'aarch64'",
  "matplotlib>=3.8.0",
  "seaborn>=0.13.0",
  "plotly>=5.20.0",
  "shap>=0.44.0",
  "lime>=0.2.0.1",
  "joblib>=1.3.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.0",
  "pyyaml>=6.0.1",
  "rich>=13.7.0",
  "click>=8.1.7",
]

[project.scripts]
pms = "pms.cli:main"

[tool.setuptools]
packages = ["pms"]

